11.4

+++ ly-common/src/main/java/com/leyou/common/vo/PageResult.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class PageResult<T> {
     private Long total;// 总条数
-    private Long totalPage;// 总页数
+    private int totalPage;// 总页数
     private List<T> items;// 当前页数据
 
     public PageResult() {
@@ -15,7 +15,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, int totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
@@ -37,11 +37,11 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public Long getTotalPage() {
+    public int getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(int totalPage) {
         this.totalPage = totalPage;
     }
